+++ lib/cov_loupe/config/option_parser_builder.rb
@@ -68,7 +68,8 @@ module CovLoupe
       end
       parser.on(
         '-f', '--format FORMAT', String,
-        'Output format: t[able]|j[son]|p[retty-json]|y[aml]|a[mazing_print] (default: table)'
+        'Output format: a[mazing_print]|i[nspect]|j[son]|J (pretty_json)|p[uts]|' \
+          'P (pretty_print)|t[able]|y[aml] (default: table)'
       ) do |value|
         config.format = normalize_format(value)
       end